Fix default image not displaying prior to load
Previously, if a default image was set, it would not be displayed
until a network load had begun. In some use cases where the load
would not be started until some arbitrary time in the future, this
would leave an empty image view visible to the user. This patch
checks to see if there is a default image set, and if so, uses that
to clear the view, so prior to loading, the defualt image will be
shown.

Change-Id: I6fcee408d95d708676d68f3f0bc7222e56570f39