Adding cache as parameter to Volley.newRequestQueue
Previously it was not possible to set the size of the cache used by the RequestQueue to anything but the default 5Mb. Adding the cache as a parameter to the Volley.newRequestQueue method opens up for setting the max size of the cache to whatever the developer wants, as well as to changing to a different implementation of the cache should that be needed.

Change-Id: Ib2f3d82d48caa023d91c1bf251c00802c7d0acc6