Extended response
It adds a flag to see if the response comes from cache or not.
If the developer interested in this flag, he/she has to overwrite
Request.deliverResponse(T response, boolean cachedResponse).

It adds also a ErrorWithCachedObjectListener.
If the developer uses this instead of ErrorListener, he/she gets
a cached result if there is one in case of an error.