Fixes null pointer when a 304 response has no cache entry.
Volley assumes that when a 304 response code is received there will be a previous response in the cache. The case where this falls short is when a
server returns a 304 response based on a query parameter instead of based on the header. In the case where the response has no content I've set the
boolean to false since the byte array provided is not actually being loaded from the cache.

Change-Id: Ic4fe8ae674870bcca199ad348d36648f677c3202
Signed-off-by: Michael Novak <michael.novakjr@gmail.com>